Proposal for temporary ginkgo v1 custom reporter
A "AddReportEntry" function, somewhat similar to ginkgo v2's one has
been implemented as part of the ptp custom reporter package.

Whatever entry added with AddReportEntry will appear always in the
system-out field of the corresponding TC.

In order to implement it, the PTP ginkgo reporter needs to be saved as a
"singleton" package variable, so the entries are saved until the xml is
generated. The, they're appended to the existing system-out text.

To use it inside any TC, just import the package and use it:

import ptpReporter
"github.com/openshift/ptp-operator/test/utils/ginkgo_reporter"

It(.. {
  ...
  ptpReporter.AddReportEntry("myEntry1", "this is the text for my
entry")
  ...
})

Whatever object can be passed as the entry value, as it will be
serialized using the stringer function (if available) or with
fmt.Sprintf("%+v", value) as fallback method.